requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
mysql-connector-python>=8.1.0
langchain>=0.0.312
langchain-openai>=0.0.1
langchain-groq>=0.0.1
openai>=1.2.4
structlog>=23.2.0
python-json-logger>=2.0.7
aiohttp>=3.8.0 